Refactor component paths
