+++ components/intermediary/request-token.tsx
@@ -1,10 +1,10 @@
 import React from 'react'
 import { Grid, GridColumn } from '@beequip/hexagon'
-import { PublicCreateIntermediary_publicCreateIntermediary_intermediary as Intermediary } from '@beequip/dev-types/graphql/PublicCreateIntermediary'
-import { CreateIntermediaryForm } from './CreateIntermediaryForm'
-import { TokenResult } from './TokenResult'
+import { PublicCreateIntermediary_publicCreateIntermediary_intermediary as Intermediary } from '@/types/graphql/PublicCreateIntermediary'
+import { CreateIntermediaryForm } from './create-intermediary-form'
+import { TokenResult } from './token-result'
 
-export const RequestToken = () => {
+export function RequestToken(): JSX.Element {
     const [intermediaryData, setIntermediaryData] = React.useState<
         Intermediary
     >()
