--- components/layout/index.tsx
@@ -1,5 +1,0 @@
-export * from './_docs'
-export * from './_introduction'
-export * from './container'
-export * from './jumbotron'
-export * from './navbar'
